Remove sqlite; temporary remove projector executor type
