+++ examples/simple_projection/src/projector.rs
@@ -1,5 +1,5 @@
 use async_trait::async_trait;
-use sqlx::{Executor, Sqlite, Transaction};
+use sqlx::{Executor, Postgres, Transaction};
 use uuid::Uuid;
 
 use esrs::store::StoreEvent;
@@ -11,23 +11,19 @@ use crate::structs::{CounterError, CounterEvent};
 pub struct CounterProjector;
 
 #[async_trait]
-impl Projector<Transaction<'static, Sqlite>, CounterEvent, CounterError> for CounterProjector {
-    async fn project(
-        &self,
-        event: &StoreEvent<CounterEvent>,
-        connection: &mut Transaction<Sqlite>,
-    ) -> Result<(), CounterError> {
-        let existing: Option<Counter> = Counter::by_id(event.aggregate_id, &mut *connection)
+impl Projector<CounterEvent, CounterError> for CounterProjector {
+    async fn project(&self, event: &StoreEvent<CounterEvent>) -> Result<(), CounterError> {
+        let existing: Option<Counter> = Counter::by_id(event.aggregate_id, todo!())
             .await?
             .map(|existing| existing);
         match event.payload {
             CounterEvent::Incremented => match existing {
-                Some(counter) => Ok(Counter::update(event.aggregate_id, counter.count + 1, &mut *connection).await?),
-                None => Ok(Counter::insert(event.aggregate_id, 1, &mut *connection).await?),
+                Some(counter) => Ok(Counter::update(event.aggregate_id, counter.count + 1, todo!()).await?),
+                None => Ok(Counter::insert(event.aggregate_id, 1, todo!()).await?),
             },
             CounterEvent::Decremented => match existing {
-                Some(counter) => Ok(Counter::update(event.aggregate_id, counter.count - 1, &mut *connection).await?),
-                None => Ok(Counter::insert(event.aggregate_id, -1, &mut *connection).await?),
+                Some(counter) => Ok(Counter::update(event.aggregate_id, counter.count - 1, todo!()).await?),
+                None => Ok(Counter::insert(event.aggregate_id, -1, todo!()).await?),
             },
         }
     }
@@ -40,7 +36,10 @@ pub struct Counter {
 }
 
 impl Counter {
-    pub async fn by_id(id: Uuid, executor: impl Executor<'_, Database = Sqlite>) -> Result<Option<Self>, sqlx::Error> {
+    pub async fn by_id(
+        id: Uuid,
+        executor: impl Executor<'_, Database = Postgres>,
+    ) -> Result<Option<Self>, sqlx::Error> {
         sqlx::query_as::<_, Self>("SELECT * FROM counters WHERE counter_id = $1")
             .bind(id)
             .fetch_optional(executor)
@@ -50,7 +49,7 @@ impl Counter {
     pub async fn insert(
         id: Uuid,
         count: i32,
-        executor: impl Executor<'_, Database = Sqlite>,
+        executor: impl Executor<'_, Database = Postgres>,
     ) -> Result<(), sqlx::Error> {
         sqlx::query_as::<_, Self>("INSERT INTO counters (counter_id, count) VALUES ($1, $2)")
             .bind(id)
@@ -63,7 +62,7 @@ impl Counter {
     pub async fn update(
         id: Uuid,
         count: i32,
-        executor: impl Executor<'_, Database = Sqlite>,
+        executor: impl Executor<'_, Database = Postgres>,
     ) -> Result<(), sqlx::Error> {
         sqlx::query_as::<_, Self>("UPDATE counters SET count = $2 WHERE counter_id = $1")
             .bind(id)
@@ -73,7 +72,7 @@ impl Counter {
             .map(|_| ())
     }
 
-    pub async fn delete(id: Uuid, executor: impl Executor<'_, Database = Sqlite>) -> Result<(), sqlx::Error> {
+    pub async fn delete(id: Uuid, executor: impl Executor<'_, Database = Postgres>) -> Result<(), sqlx::Error> {
         sqlx::query_as::<_, Self>("DELETE FROM counter WHERE counter_id = $1")
             .bind(id)
             .fetch_optional(executor)
