+++ examples/simple_side_effect/src/aggregate.rs
@@ -1,10 +1,10 @@
 use async_trait::async_trait;
-use sqlx::{Pool, Sqlite, Transaction};
+use sqlx::{Pool, Postgres};
 
 use esrs::aggregate::{Aggregate, AggregateManager, AggregateState};
 use esrs::policy::Policy;
 use esrs::projector::Projector;
-use esrs::store::{EventStore, SqliteStore, StoreEvent};
+use esrs::store::{EventStore, PgStore, StoreEvent};
 
 use crate::structs::{LoggingCommand, LoggingError, LoggingEvent};
 
@@ -15,8 +15,8 @@ use crate::structs::{LoggingCommand, LoggingError, LoggingEvent};
 pub struct LoggingProjector;
 
 #[async_trait]
-impl Projector<Transaction<'static, Sqlite>, LoggingEvent, LoggingError> for LoggingProjector {
-    async fn project(&self, event: &StoreEvent<LoggingEvent>, _: &mut Transaction<Sqlite>) -> Result<(), LoggingError> {
+impl Projector<LoggingEvent, LoggingError> for LoggingProjector {
+    async fn project(&self, event: &StoreEvent<LoggingEvent>) -> Result<(), LoggingError> {
         let id = event.aggregate_id;
         match event.payload() {
             LoggingEvent::Logged(msg) => {
@@ -40,8 +40,8 @@ impl Projector<Transaction<'static, Sqlite>, LoggingEvent, LoggingError> for Log
 pub struct LoggingPolicy;
 
 #[async_trait]
-impl Policy<Pool<Sqlite>, LoggingEvent, LoggingError> for LoggingPolicy {
-    async fn handle_event(&self, event: &StoreEvent<LoggingEvent>, _: &Pool<Sqlite>) -> Result<(), LoggingError> {
+impl Policy<LoggingEvent, LoggingError> for LoggingPolicy {
+    async fn handle_event(&self, event: &StoreEvent<LoggingEvent>) -> Result<(), LoggingError> {
         let id = event.aggregate_id;
         match event.payload() {
             LoggingEvent::Logged(msg) => {
@@ -56,27 +56,26 @@ impl Policy<Pool<Sqlite>, LoggingEvent, LoggingError> for LoggingPolicy {
 }
 
 // A store of events
-pub type LogStore = SqliteStore<LoggingEvent, LoggingError>;
+pub type LogStore = PgStore<LoggingEvent, LoggingError>;
 
 pub struct LoggingAggregate {
     event_store: LogStore,
 }
 
 impl LoggingAggregate {
-    pub async fn new(pool: &Pool<Sqlite>) -> Result<Self, LoggingError> {
+    pub async fn new(pool: &Pool<Postgres>) -> Result<Self, LoggingError> {
         Ok(Self {
             event_store: Self::new_store(pool).await?,
         })
     }
 
-    async fn new_store(pool: &Pool<Sqlite>) -> Result<LogStore, LoggingError> {
-        let projectors: Vec<Box<dyn Projector<Transaction<Sqlite>, LoggingEvent, LoggingError> + Send + Sync>> =
+    async fn new_store(pool: &Pool<Postgres>) -> Result<LogStore, LoggingError> {
+        let projectors: Vec<Box<dyn Projector<LoggingEvent, LoggingError> + Send + Sync>> =
             vec![Box::new(LoggingProjector)];
 
-        let policies: Vec<Box<dyn Policy<Pool<Sqlite>, LoggingEvent, LoggingError> + Send + Sync>> =
-            vec![Box::new(LoggingPolicy)];
+        let policies: Vec<Box<dyn Policy<LoggingEvent, LoggingError> + Send + Sync>> = vec![Box::new(LoggingPolicy)];
 
-        SqliteStore::new::<Self>(pool, projectors, policies).await
+        PgStore::new::<Self>(pool, projectors, policies).await
     }
 }
