+++ src/esrs/event.rs
@@ -7,7 +7,7 @@ use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use sqlx::types::Json;
-use sqlx::{Executor, Postgres, Sqlite};
+use sqlx::{Executor, Postgres};
 use uuid::Uuid;
 
 use crate::esrs::store::StoreEvent;
@@ -36,7 +36,7 @@ impl<E: Serialize + DeserializeOwned + Send + Sync> TryInto<StoreEvent<E>> for E
     }
 }
 
-#[cfg(any(feature = "postgres", feature = "sqlite"))]
+#[cfg(feature = "postgres")]
 pub trait Querier<Database: sqlx::Database> {
     fn insert<'a, Evt, Err>(
         aggregate_name: &'a str,
@@ -143,76 +143,6 @@ impl Querier<Postgres> for Event {
     }
 }
 
-#[cfg(feature = "sqlite")]
-impl Querier<Sqlite> for Event {
-    fn insert<'a, Evt, Err>(
-        aggregate_name: &'a str,
-        event_id: Uuid,
-        aggregate_id: Uuid,
-        event: &'a Evt,
-        occurred_on: DateTime<Utc>,
-        sequence_number: SequenceNumber,
-        executor: impl Executor<'a, Database = Sqlite> + 'a,
-    ) -> Pin<Box<dyn Future<Output = Result<(), Err>> + Send + 'a>>
-    where
-        Self: Sync + 'a,
-        Evt: Serialize + DeserializeOwned + Send + Sync,
-        Err: From<sqlx::Error> + From<serde_json::Error> + Send + Sync,
-    {
-        Box::pin(async move {
-            Ok(sqlx::query(insert_query(aggregate_name).as_str())
-                .bind(event_id)
-                .bind(aggregate_id)
-                .bind(Json(event))
-                .bind(occurred_on)
-                .bind(sequence_number)
-                .execute(executor)
-                .await
-                .map(|_| ())?)
-        })
-    }
-
-    fn by_aggregate_id<'a, Evt, Err>(
-        aggregate_name: &'a str,
-        aggregate_id: Uuid,
-        executor: impl Executor<'a, Database = Sqlite> + 'a,
-    ) -> Pin<Box<dyn Future<Output = Result<Vec<StoreEvent<Evt>>, Err>> + Send + 'a>>
-    where
-        Self: Sync + 'a,
-        Evt: Serialize + DeserializeOwned + Send + Sync,
-        Err: From<sqlx::Error> + From<serde_json::Error> + Send + Sync,
-    {
-        Box::pin(async move {
-            sqlx::query_as::<_, Self>(by_aggregate_id_query(aggregate_name).as_str())
-                .bind(aggregate_id)
-                .fetch_all(executor)
-                .await?
-                .into_iter()
-                .map(|event| Ok(event.try_into()?))
-                .collect::<Result<Vec<StoreEvent<Evt>>, Err>>()
-        })
-    }
-
-    fn delete_by_aggregate_id<'a, Evt, Err>(
-        aggregate_name: &'a str,
-        aggregate_id: Uuid,
-        executor: impl Executor<'a, Database = Sqlite> + 'a,
-    ) -> Pin<Box<dyn Future<Output = Result<(), Err>> + Send + 'a>>
-    where
-        Self: Sync + 'a,
-        Evt: Serialize + DeserializeOwned + Send + Sync,
-        Err: From<sqlx::Error> + From<serde_json::Error> + Send + Sync,
-    {
-        Box::pin(async move {
-            Ok(sqlx::query(delete_query(aggregate_name).as_str())
-                .bind(aggregate_id)
-                .execute(executor)
-                .await
-                .map(|_| ())?)
-        })
-    }
-}
-
 fn by_aggregate_id_query(table_name: &str) -> String {
     format!(
         "SELECT * FROM {}_events WHERE aggregate_id = $1 ORDER BY sequence_number ASC",
@@ -235,6 +165,6 @@ fn delete_query(aggregate_name: &str) -> String {
     format!("DELETE FROM {}_events WHERE aggregate_id = $1", aggregate_name)
 }
 
-pub fn select_all_query(aggregate_name: &str) -> String {
-    format!("SELECT * FROM {}_events ORDER BY sequence_number ASC", aggregate_name)
-}
+// pub fn select_all_query(aggregate_name: &str) -> String {
+//     format!("SELECT * FROM {}_events ORDER BY sequence_number ASC", aggregate_name)
+// }
