+++ src/esrs/policy.rs
@@ -5,9 +5,9 @@ use serde::Serialize;
 use crate::esrs::store::StoreEvent;
 
 #[async_trait]
-pub trait Policy<Executor, Event: Serialize + DeserializeOwned + Send + Sync, Error> {
+pub trait Policy<Event: Serialize + DeserializeOwned + Send + Sync, Error> {
     /// This function intercepts the event and, matching on the type of such event
     /// produces the appropriate side effects.
     /// The result is meant to catch generic errors.
-    async fn handle_event(&self, event: &StoreEvent<Event>, executor: &Executor) -> Result<(), Error>;
+    async fn handle_event(&self, event: &StoreEvent<Event>) -> Result<(), Error>;
 }
